@@ -1,11 +1,11 @@
 name: pigeon
-version: 0.1.24 # This must match the version in lib/generator_tools.dart
+version: 0.2.0 # This must match the version in lib/generator_tools.dart
 description: Code generator tool to make communication between Flutter and the host platform type-safe and easier.
 homepage: https://github.com/flutter/packages/tree/master/packages/pigeon
 dependencies:
-  args: ^1.5.2
-  path: ^1.6.4
+  args: ^2.0.0
+  path: ^1.8.0
 dev_dependencies:
   test: ^1.11.1
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
