@@ -5,18 +5,18 @@
 import 'package:pigeon/pigeon.dart';
 
 class Everything {
-  bool aBool;
-  int anInt;
-  double aDouble;
-  String aString;
-  Uint8List aByteArray;
-  Int32List a4ByteArray;
-  Int64List a8ByteArray;
-  Float64List aFloatArray;
+  bool? aBool;
+  int? anInt;
+  double? aDouble;
+  String? aString;
+  Uint8List? aByteArray;
+  Int32List? a4ByteArray;
+  Int64List? a8ByteArray;
+  Float64List? aFloatArray;
   // ignore: always_specify_types
-  List aList;
+  List? aList;
   // ignore: always_specify_types
-  Map aMap;
+  Map? aMap;
 }
 
 @HostApi()
